+++ logic/skill.lua
@@ -169,22 +169,19 @@ MTSL_LOGIC_SKILL = {
         return is_obtainable
     end,
 
-    ------------------------------------------------------------------------------------------------
-    -- Returns a skill for a certain profession based on the recipe id its learned from
-    --
-    -- @item_id         Number      The id of the item (source of the skill)
-    -- @prof_name       String      Name of the profession
-    --
-    -- returns          Array       The skills
-    ------------------------------------------------------------------------------------------------
-    GetSkillForProfessionByItemId = function(self, item_id, profession_name)
-        local skill = MTSL_TOOLS:GetItemFromArrayByKeyArrayValue(TRADE_SKILLS_DATA["skills"][profession_name], "items", item_id)
-        -- try a level if nil
-        if skill == nil then
-            skill = MTSL_TOOLS:GetItemFromArrayByKeyArrayValue(TRADE_SKILLS_DATA["levels"][profession_name], "items", item_id)
-        end
-        return skill
-    end,
+    GetSkillForProfessionByItemId =
+        ---
+        --- Returns a skill for a certain profession based on the recipe id its learned from
+        ---@param item_id number
+        ---@param profession_name string
+        ---@return Skill|nil
+        function(self, item_id, profession_name)
+            local skill = MTSL_TOOLS:GetItemFromArrayByKeyArrayValue(TRADE_SKILLS_DATA["skills"][profession_name], "items", item_id)
+            if skill == nil then
+                skill = MTSL_TOOLS:GetItemFromArrayByKeyArrayValue(TRADE_SKILLS_DATA["levels"][profession_name], "items", item_id)
+            end
+            return skill
+        end,
 
     ------------------------------------------------------------------------------------------------
     -- Returns a skill for a certain profession by id