+++ ui/tooltip.lua
@@ -3,79 +3,78 @@
 ----------------
 local _G = _G or getfenv(0)
 
-MTSL_RECIPE_PROFESSION_TYPES  = {
-    "Leatherworking",   -- _G.LE_ITEM_RECIPE_LEATHERWORKING    = 1
-    "Tailoring",        -- _G.LE_ITEM_RECIPE_TAILORING         = 2
-    "Engineering",      -- _G.LE_ITEM_RECIPE_ENGINEERING       = 3
-    "Blacksmithing",    -- _G.LE_ITEM_RECIPE_BLACKSMITHING     = 4
-    "Cooking",          -- _G.LE_ITEM_RECIPE_COOKING           = 5
-    "Alchemy",          -- _G.LE_ITEM_RECIPE_ALCHEMY           = 6
-    "First Aid",        -- _G.LE_ITEM_RECIPE_FIRST_AID         = 7
-    "Enchanting",       -- _G.LE_ITEM_RECIPE_ENCHANTING        = 8
-}
+---@param tooltip GameTooltip
+---@param link string
+---@param text string
+---@return boolean|nil
+local function enhanceTooltip(tooltip, link, text)
+    local _, _, string_item_id = strfind(link, 'item:(%d+)')
+    if string_item_id == nil then
+        return false
+    end
 
--- variable to avoid adding the info more than once to the tooltip
-local MTSL_TOOLTIP_SHOWN = 0
+    ---@type number
+    local item_id = --[[---@type number]] tonumber(--[[---@type string]] string_item_id)
+    if item_id == nil then
+        return false
+    end
 
-if false then
--- Enrich the tooltip with alts info
-GameTooltip:HookScript("OnTooltipSetItem", function(self)
-    -- Only update the tooltip when needed
-    if MTSLUI_SAVED_VARIABLES:GetEnhancedTooltipActive() and MTSL_TOOLTIP_SHOWN <= 0 then
-         -- Mark the tooltip as shown
-        MTSL_TOOLTIP_SHOWN = 1
-        -- Find out if this is the first or second call of OnTooltipSetItem().
-        local name, link = GameTooltip:GetItem()
-        if not name or not link then return end
-        -- Get the type of item we are gonna show in the tooltip
-        local _, _, _, _, _, _, _, _, _, _, _, itemTypeId, subItemTypeId = GetItemInfo(link)
-        -- Only looking at recipes/formula's/schematics/...4
-        if itemTypeId == _G.LE_ITEM_CLASS_RECIPE then
-            local _, itemId, _  = strsplit(":", link)
-            local prof_name = MTSL_RECIPE_PROFESSION_TYPES[subItemTypeId]
-            -- add a line for each character on the same realm (but not current player itself) that knows the profession
-            local other_players
-            if MTSLUI_SAVED_VARIABLES:GetEnhancedTooltipFaction() == "any" then
-                other_players = MTSL_LOGIC_PLAYER_NPC:GetOtherPlayersOnCurrentRealmLearnedProfession(prof_name)
+    local _, _, _, _, item_type, item_sub_type, _, _, _ = GetItemInfo(item_id)
+    if item_type ~= "Recipe" then
+        return false
+    end
+
+    local profession = item_sub_type
+
+    local other_players
+    if MTSLUI_SAVED_VARIABLES:GetEnhancedTooltipFaction() == "any" then
+        other_players = MTSL_LOGIC_PLAYER_NPC:GetOtherPlayersOnCurrentRealmLearnedProfession(profession)
+    else
+        other_players = MTSL_LOGIC_PLAYER_NPC:GetOtherPlayersOnCurrentRealmSameFactionLearnedProfession(profession)
+    end
+    if MTSL_TOOLS:CountItemsInArray(other_players) == 0 then
+        return false
+    end
+
+    local skill = MTSL_LOGIC_SKILL:GetSkillForProfessionByItemId(item_id, profession)
+    if skill == nil then
+        return false
+    end
+
+    tooltip:AddLine(" ")
+    for _, player in pairs(other_players) do
+        local player_profession = player["TRADESKILLS"][profession]
+
+        local status = "known"
+        local status_color = MTSLUI_FONTS.COLORS.AVAILABLE.YES
+        local description = "Already known by"
+        if MTSL_TOOLS:ListContainsNumber(player_profession["MISSING_SKILLS"], tonumber(skill.id)) == true then
+            if tonumber(player_profession["SKILL_LEVEL"]) >= tonumber(skill.min_skill) then
+                status = "learnable"
+                status_color = MTSLUI_FONTS.COLORS.AVAILABLE.LEARNABLE
+                description = "Could be learned by"
             else
-                other_players = MTSL_LOGIC_PLAYER_NPC:GetOtherPlayersOnCurrentRealmSameFactionLearnedProfession(prof_name)
-            end
-
-            -- Only add if we have players to add
-            if MTSL_TOOLS:CountItemsInArray(other_players) > 0 then
-                -- Get the skill from our data (convert itemid to a number instead of string) so we can check the min_skill
-                local skill = MTSL_LOGIC_SKILL:GetSkillForProfessionByItemId(tonumber(itemId), prof_name)
-                if skill ~= nil then
-                    -- Empty line
-                    GameTooltip:AddLine(" ")
-                    GameTooltip:AddLine(MTSLUI_TOOLS:GetLocalisedLabel("status other characters"), true)
-                    for _, v in pairs(other_players) do
-                        -- Check if learned or not
-                        -- default color = learned (green)
-                        local status_color = MTSLUI_FONTS.COLORS.AVAILABLE.YES
-                        if MTSL_TOOLS:ListContainsNumber(v["TRADESKILLS"][prof_name]["MISSING_SKILLS"], tonumber(skill.id)) == true then
-                            -- depending on his skill level, mark as learnable or not yet learnable
-                            if tonumber(v["TRADESKILLS"][prof_name]["SKILL_LEVEL"]) >= tonumber(skill.min_skill) then
-                                status_color = MTSLUI_FONTS.COLORS.AVAILABLE.LEARNABLE
-                            else
-                                status_color = MTSLUI_FONTS.COLORS.AVAILABLE.NO
-                            end
-                        end
-                        -- dont add to tooltip if we hide known players
-                        if (MTSLUI_SAVED_VARIABLES:GetEnhancedTooltipShowKnown() and status_color == MTSLUI_FONTS.COLORS.AVAILABLE.YES) or
-                                status_color ~= MTSLUI_FONTS.COLORS.AVAILABLE.YES then
-                            local faction_color = MTSLUI_FONTS.COLORS.FACTION[string.upper(v.FACTION)]
-                            GameTooltip:AddLine(MTSLUI_FONTS.TAB .. status_color  .. "[" .. v["TRADESKILLS"][prof_name]["SKILL_LEVEL"] .. "] " .. faction_color .. v["NAME"])
-                        end
-                    end
-                end
+                status = "no"
+                status_color = MTSLUI_FONTS.COLORS.AVAILABLE.NO
+                description = "Will be learnable by"
             end
         end
-    end
-end)
 
--- Mark the tooltip as hidden
-GameTooltip:HookScript("OnTooltipCleared", function(self)
-    MTSL_TOOLTIP_SHOWN = 0
-end)
+        if status ~= "known" or MTSLUI_SAVED_VARIABLES:GetEnhancedTooltipShowKnown() then
+            local faction_color = MTSLUI_FONTS.COLORS.FACTION[string.upper(player.FACTION)]
+            tooltip:AddLine(status_color .. description .. " " .. faction_color .. player["NAME"] .. status_color .. " " .. "(" .. player_profession["SKILL_LEVEL"] .. ") ")
+        end
+    end
+    tooltip:Show()
+
+    return true
+end
+
+local MTSL_HookSetItemRef = SetItemRef
+---@param link string
+---@param text string
+---@param button MouseButton
+SetItemRef = function(link, text, button)
+    MTSL_HookSetItemRef(link, text, button)
+    enhanceTooltip(ItemRefTooltip, link, text)
 end